Refactor: 일정 삭제 기능 리팩토링
- Controller 레이어
> - deleteSchedule 매개변수로 HttpServletRequest 추가
> - 서블릿의 getAttribute 를 통해 userId 값 얻은 후
> - 서비스 레이어에 요청 DTO 와 함께 전달

- Service 레이어
> - 매개변수로 받은 userId(Session 활용하여 얻은 값) 와 수정하려는 일정 데이터의 userId와 비교
> - 일치: 정상 삭제, 불일치: 403 FORBIDDEN

+++ src/main/java/com/example/schedulerapp/service/ScheduleServiceJPA.java
@@ -80,8 +80,12 @@ public class ScheduleServiceJPA implements ScheduleService {
     }
 
     @Override
-    public void deleteSchedule(Long id) {
-        Schedule findSchedule = scheduleRepository.findByIdOrElseThrow(id);
+    public void deleteSchedule(Long scheduleId, Long userId) {
+        Schedule findSchedule = scheduleRepository.findByIdOrElseThrow(scheduleId);
+
+        if (!findSchedule.getUser().getId().equals(userId)) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Only the author of the schedule can delete it.");
+        }
 
         scheduleRepository.delete(findSchedule);
     }
